Test reflecting that PEP550 semantics "just work"
Picking an example from PEP 550, where local context
inside a generator is expected to work, shielding
local changes and exposing changes in unmodified
variables.

There are variations which don't work
with the "native" backend, reflected
in a test marked to be skipped.

Also: bumps version and development status to "production".